Move some drawing-specific code into the appropriate directory
